requirements.txt
aiohttp
requests
python-dotenv
solana
base58
python-telegram-bot==21.10
nest-asyncio